Crate pallet_nftdao

NFTDAO Module

A pallet that refers to Moloch and can use NFT as tribute.

Terminology

Pool: It can be exchanged with some FTs, and the price can be automatically discovered through bancor curve.

Action: After the proposal is passed, the operation of the dao account on the chain.

Proposal Queue: Only proposals in the queue can be voted.

Sponsor: In order to prevent spam proposals, a proposal must be sponsored to enter the queue.

Vote: Yes or not, only members of dao can vote.

Grace: You can ragequit after the voting period.

Ragequit: Burn shares and exchange for corresponding assets.

Interface

Dispatchable Functions

create_dao - Create a new DAO.

submit_proposal - Submit a proposal, regardless of whether it is a member of dao can perform this operation.

cancel_proposal - The proposal can be cancelled before it is sponsored.

sponsor_proposal - Sponsor a proposal and make it into the queue.

vote_proposal - DAO members can vote on proposals.

process_proposal - After the grace period, the proposal needs to be processed.

ragequit - Burn shares and exchange for corresponding assets..

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

DAOId

DAOInfo
DAO’s details

DAOs
A set of daos. dao account => dao info

Escrows
A set of dao’s escrow account. dao account => dao escrows account

LastProposalId
The last proposalId in each DAO. dao account => proposal id

LastQueueIndex
The last proposal queue index in each DAO. dao account => proposal in queue index

Member
Member’s details

Members
A set of members in the DAO. (dao account, member account) => member

Module

Proposal
Proposal’s details

ProposalQueues
Proposal ids in the queue. (dao account, proposal queue index) => proposal id

Proposals
A set of proposals. (dao account, proposal id) => proposal

VoteMembers
A set of members who voted on the proposal (dao account, proposal queue index), member account => ()

Enums

Call
Dispatchable calls.

Error
Errors inform users that something went wrong.

RawEvent
Events for this module.

Traits

Config
Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

Event
RawEvent specialized for the configuration Config