Struct pallet_nftdao::Module

pub struct Module<T: Config>(_);

Implementations

impl<T: Config + 'static> Module<T>

pub fn get_nonce() -> u128

A self-increasing number used to create a dao account.

pub fn dao<K: EncodeLike<T::AccountId>>(
key: K
) -> DAOInfo<T::AccountId, T::BlockNumber, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>

A set of daos. dao account => dao info

pub fn member<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Member where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::AccountId>,

pub fn escrow<K: EncodeLike<T::AccountId>>(key: K) -> T::AccountId

A set of dao’s escrow account. dao account => dao escrows account

pub fn last_proposal_id<K: EncodeLike<T::AccountId>>(key: K) -> Option<u128>

The last proposalId in each DAO. dao account => proposal id

pub fn proposal<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Option<Proposal<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, T::Hash>> where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<u128>,

pub fn last_queue_index<K: EncodeLike<T::AccountId>>(key: K) -> Option<u128>

The last proposal queue index in each DAO. dao account => proposal in queue index

pub fn proposal_queue<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> u128 where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<u128>,

pub fn vote_member<KArg1, KArg2>(k1: KArg1, k2: KArg2) where
KArg1: EncodeLike<(T::AccountId, u128)>,
KArg2: EncodeLike<T::AccountId>,

impl<T: Config> Module<T>

Can also be called using Call.

pub fn create_dao(
origin: T::Origin,
metadata: Vec<u8>,
period_duration: u128,
voting_period: u128,
grace_period: u128,
shares_requested: u128,
proposal_deposit: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
processing_reward: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
dilution_bound: u128
) -> DispatchResult

Create a new DAO.

The dispatch origin of this call must be Signed.

Parameters:

metadata: The details of DAO, can be a ipfs CID.

period_duration: The duration of each period.

voting_period: How many periods does voting last.

grace_period: How many periods does grace last.

shares_requested: How many shares does the summoner need.

shares_requested: How many shares does the summoner need.

proposal_deposit: How many asset does the sponner need.

processing_reward: How much is the reward for processing the proposal.

dilution_bound: Value that protects members from extreme dilution, default is 3.

NOTE: Calling this function will bypass origin filters.

pub fn submit_proposal(
origin: T::Origin,
dao_account: T::AccountId,
applicant: T::AccountId,
shares_requested: u128,
tribute_offered: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
tribute_nft: Option<(T::Hash, u128)>,
details: Vec<u8>,
action: Option<Vec<u8>>
) -> DispatchResult

Submit a new proposal.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: Account for submitting proposals to that dao.

applicant: Who will act after the proposal is passed.

shares_requested: Share to applicant after the proposal passed.

tribute_offered: Tribute to dao from the person who submitted the proposal.

tribute_nft: Tribute NFT to dao from the person who submitted the proposal.

details: The detailed information of the proposal, which can be ipfs cid with title and description.

action: Operations performed by DAO, such as transfers or purchases.

NOTE: Calling this function will bypass origin filters.

pub fn cancel_proposal(
origin: T::Origin,
dao_account: T::AccountId,
proposal_id: u128
) -> DispatchResult

Cancel the proposal and return the tribute.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: The account of the dao where the proposal is to be cancelled.

proposal_id: The id of proposal.

NOTE: Calling this function will bypass origin filters.

pub fn sponsor_proposal(
origin: T::Origin,
dao_account: T::AccountId,
proposal_id: u128
) -> DispatchResult

Sponsor a proposal and make it into the queue.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: The account of the dao where the proposal is to be sponsored.

proposal_id: The id of proposal.

NOTE: Calling this function will bypass origin filters.

pub fn vote_proposal(
origin: T::Origin,
dao_account: T::AccountId,
proposal_index: u128,
yes: bool
) -> DispatchResult

Vote for the proposal.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: The account of the dao where the proposal is to be voted.

proposal_index: Proposal indexing in the queue.

yes: true is yes, false is not

NOTE: Calling this function will bypass origin filters.

pub fn process_proposal(
origin: T::Origin,
dao_account: T::AccountId,
proposal_index: u128
) -> DispatchResult

Processing proposal.

If the proposal fails, the tribute will be returned. After processing the proposal, part of the funds sponsored by the sponsor will be given to the processor as a reward, and part will be returned to the sponsor.

If the proposal includes an action, ProposalExecuted event will be emitted after the action is executed. The ProposalProcessed event will be emitted after the proposal is processed.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: The account of the dao where the proposal is to be processed.

proposal_index: Proposal indexing in the queue.

NOTE: Calling this function will bypass origin filters.

pub fn ragequit(
origin: T::Origin,
dao_account: T::AccountId,
shares_to_burn: u128
) -> DispatchResult

Members can burn shares and obtain corresponding assets.

The dispatch origin of this call must be Signed.

Parameters:

dao_account: The account of the dao where the proposal is to be ragequited.

shares_to_burn: How many shares to burn.

NOTE: Calling this function will bypass origin filters.

impl<T: Config> Module<T>

pub fn account_id() -> T::AccountId

Account of this pallet.

pub fn dao_id(
summoner_address: &T::AccountId,
details: &[u8]
) -> Result<DAOId, DispatchError>

Convert the hash value to DAOId.

pub fn dao_account_id(dao_id: &DAOId) -> T::AccountId

Convert the DAOId to dao account.

pub fn dao_escrow_id(dao_id: &DAOId) -> T::AccountId

Use dao account to generate escrow id (a sub account).

pub fn proposal_id_increment(
dao_account: &T::AccountId
) -> Result<u128, DispatchError>

id plus one

pub fn queue_index_increment(
dao_account: &T::AccountId
) -> Result<u128, DispatchError>

index plus one

pub fn get_current_period(
dao: &DAOInfo<T::AccountId, T::BlockNumber, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>
) -> Result<u128, DispatchError>

dao in the period of the current block

pub fn calculate_starting_period(
dao: &DAOInfo<T::AccountId, T::BlockNumber, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>
) -> Result<u128, DispatchError>

Calculate the period for the proposal to start voting

pub fn add_vote(now_vote: u128, shares: u128) -> Result<u128, DispatchError>

increase the number of votes with shares.

pub fn run(
dao_account: T::AccountId,
action_data: &[u8]
) -> Result<bool, DispatchError>

Let dao perform an operation (call).